Also move device
