+++ vllm/utils/__init__.py
@@ -67,11 +67,6 @@ STR_INVALID_VAL: str = "INVALID"
 T = TypeVar("T")
 
 
-class Device(enum.Enum):
-    GPU = enum.auto()
-    CPU = enum.auto()
-
-
 class LayerBlockType(enum.Enum):
     attention = "attention"
     mamba = "mamba"
